Database Updating

+++ TunnelVision/app/src/main/java/com/way/tunnelvision/entity/dao/DaoMaster.java
@@ -5,6 +5,11 @@ import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteDatabase.CursorFactory;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.util.Log;
+
+import com.way.tunnelvision.base.Constants;
+import com.way.tunnelvision.base.MainApp;
+import com.way.tunnelvision.util.PreferencesUtil;
+
 import de.greenrobot.dao.AbstractDaoMaster;
 import de.greenrobot.dao.identityscope.IdentityScopeType;
 
@@ -12,30 +17,35 @@ import de.greenrobot.dao.identityscope.IdentityScopeType;
  * Created by pc on 2016/2/15.
  * Master of DAO (schema version 1000): knows all DAOs.
  */
-public class DaoMaster  extends AbstractDaoMaster {
+public class DaoMaster extends AbstractDaoMaster {
     private final static String TAG = DaoMaster.class.getName();
 
     public static final int SCHEMA_VERSION = 1000;
+    public static final int SCHEMA_VERSION_NEW = 1001;
 
     public DaoMaster(SQLiteDatabase db) {
         super(db, SCHEMA_VERSION);
         registerDaoClass(MenuDao.class);
         registerDaoClass(ChannelDao.class);
-        //registerDaoClass(OrderDao.class);
+        registerDaoClass(NewsDetailDao.class);
     }
 
-    /** Creates underlying database table using DAOs. */
+    /**
+     * Creates underlying database table using DAOs.
+     */
     public static void createAllTables(SQLiteDatabase db, boolean ifNotExists) {
         MenuDao.createTable(db, ifNotExists);
         ChannelDao.createTable(db, ifNotExists);
-//        OrderDao.createTable(db, ifNotExists);
+        NewsDetailDao.createTable(db, ifNotExists);
     }
 
-    /** Drops underlying database table using DAOs. */
+    /**
+     * Drops underlying database table using DAOs.
+     */
     public static void dropAllTables(SQLiteDatabase db, boolean ifExists) {
         MenuDao.dropTable(db, ifExists);
         ChannelDao.dropTable(db, ifExists);
-//        OrderDao.dropTable(db, ifExists);
+        NewsDetailDao.dropTable(db, ifExists);
     }
 
     public static abstract class OpenHelper extends SQLiteOpenHelper {
@@ -48,10 +58,13 @@ public class DaoMaster  extends AbstractDaoMaster {
         public void onCreate(SQLiteDatabase db) {
             Log.i(TAG, "Creating tables for schema version " + SCHEMA_VERSION);
             createAllTables(db, false);
+            PreferencesUtil.putInt(MainApp.getContext(), Constants.PREFERENCE_KEY_DATABASE_VERSION_CURRENT, SCHEMA_VERSION);
         }
     }
 
-    /** WARNING: Drops all table on Upgrade! Use only during development. */
+    /**
+     * WARNING: Drops all table on Upgrade! Use only during development.
+     */
     public static class DevOpenHelper extends OpenHelper {
         public DevOpenHelper(Context context, String name, CursorFactory factory) {
             super(context, name, factory);
@@ -59,9 +72,12 @@ public class DaoMaster  extends AbstractDaoMaster {
 
         @Override
         public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-            Log.i(TAG, "Upgrading schema from version " + oldVersion + " to " + newVersion + " by dropping all tables");
-            dropAllTables(db, true);
-            onCreate(db);
+            if (newVersion > oldVersion) {
+                Log.i(TAG, "Upgrading schema from version " + oldVersion + " to " + newVersion + " by dropping all tables");
+                dropAllTables(db, true);
+                onCreate(db);
+                PreferencesUtil.putInt(MainApp.getContext(), Constants.PREFERENCE_KEY_DATABASE_VERSION_CURRENT, newVersion);
+            }
         }
     }
